Modify path in ingestion to catch descendants
Modify the pattern as we ingest it to add ** on either
side as applicable, so that we match children. This
removes the need for many of the "if" match statements.
One possible issue is that in the case of an exclusion,
it appears that now the directory does not match
but its children then will; this generates the same
behavior as far as the tests are concerned. This is noticed
in that .terraform/modules gets "excluded" but then its
children are included, so there is no problem, but it
looks iffy in the debug logs