Add terraform.tfstate to default exclusion list
TFE/TFC will return an error if `terraform.tfstate` is detected in the
slug that is uploaded to the remote agent. In TFE/TFC we read state from
the cloud rather than a state file, so we do not need
`terraform.tfstate` to be included in the slug.

This commit adds `terraform.tfstate` to the default exclusions list,
which is used if, and only if, there is no `.terraformignore` file
specified in the project directory. So if that file is included the
`terraform.tfstate` file WILL be included in the slug, unless of course
it is added to the `.terraformignore` file.

This gives us flexibility to have the default state file included if
needed and excluded by default.