slugoverlay package
This package is a composition of go-slug that can be used to generate an overlay filesystem archive that can preserve file system operations on an identical directory that it was based on. This can be useful when processing a large filesystem and storing only what has changed in order to re-create the changed filesystem at a later time.

It works by storing a crc32 checksum for each file in a target directory, then comparing the files against that basis during packing. File deletions are preserved using .tombstone metadata files within the archive, so slugs created using PackOverlay should be unpacked using UnpackOverlay despite being compatible with go-slug Unpack.

All of the existing Packer options should be compatible with this methodology. A new package was defined in order to maintain long term independence and compatibility with the nominal slug package.