ignorefiles: .terraformignore handling in a separate package
The handling of .terraformignore files was previously just embedded
directly in the root package of this repository as a bunch of unexported
symbols.

In future commits we'll introduce another package to this codebase that
will also need to deal with .terraformignore files, but in a slightly
different context. In preparation, here we factor out that handling into
a separate internal package and adapt the root package to call into that
package instead of using its own inline implementation.

This carefully preserves the existing caller's ability to ignore various
errors, while now allowing new callers to handle errors differently.