sourceaddrs: Parsers and models for source package addresses
This package models the source address types we'll support for modules
when working with the forthcoming "source bundle" mechanism, which is
essentially a "meta-slug" capturing the contents of many different source
packages all at once.

For the resulting source bundle to be useful we need to be able to talk
about the source addresses it was built from, so that callers can request
the bundle path equivalent to a particular remote source address path.

This address syntax is intentionally a subset of the go-getter-based
syntax used in today's Terraform CLI module installer, because we want
to have a set of commonly-used address types that are treated equivalently
in both implementations. However, we won't be actually using go-getter
here because its high level of flexibility (and quirkiness) is a poor fit
for safely constructing persistable source code bundles.