sourceaddrs: SourceFilename helper function
We typically decide between using HCL native syntax vs. JSON HCL syntax
based on filename suffixes, and so it's helpful to have a robust way to
retrieve just the filename portion of an arbitrary source address (as long
as it's referring to a file) which takes into account edge-cases such as
the query string portion of remote source addresses.