sourceaddrs: FinalSource interface type
Terraform module source addresses have a historical design oddity where
both local and remote source addresses are self-contained but registry
source addresses must always be combined with a version constraint and
then resolved through an extra step of selecting the latest available
version that matches that version constraint.

This means that we cannot use the same type to represent both a package
to be installed and a package that has already been resolved. To fill that
gap here we introduce a variant address type sourceaddrs.FinalSource which
represents addresses of packages that have already been installed, rather
that addresses of packages that are requested to be installed.

LocalSource and RemoteSource implement both interfaces, while
RegistrySource only implements Source and the new RegistrySourceFinal
only implements FinalSource. This then allows other subsystems that are
built in terms of these address types to be explicit about whether they
are expecting an address of something to be installed or an address of
something that has already been installed, and thus the Go compiler can
help ensure that we handle both cases fully rather than forgetting about
the need for the extra version selection information for already-installed
registry module source addresses.