sourcebundle and sourceaddrs: Both are experimental
Since this is entirely new functionality that wasn't previously in this
codebase's scope, we'll need to gather some real experience using this new
API before we can be confident that it's suitable to commit to.

At the time of this commit this entire codebase is pre-1.0 and so subject
to breaking changes anyway, but these extra notes are here just in case
the root package functionality gets stabilized before these new packages
are ready to be stabilized. We'll remove these warnings at a later date
once we've got enough experience with these new packages in internal
projects to confidently commit to a public-facing API.