sourceaddrs: Use correct canonical form for current directory
We previously missed an exception for making the canonical form of a
local source referring to the current directory be "./", to match the
parent directory form "../". This logic was incorrectly requiring that
it be written as "./.", which is silly.

Now "./" and "../" are both accepted as canonical forms, and naked "." and
".." are still rejected as non-canonical as before, which we do just to
avoid injecting further ambiguity into this already-rather-ambiguous
source address syntax so that we'll have more options for potentially
expanding this microsyntax in future.