Static defaultExclusions rules were not congruent
The default rule set did not reflect the necessary "negationsAfter"
state for rules that appear before the default negation rule.

This caused Pack to skip the entire ".terraform/modules" directory
when it was encountered because it believed the rule to be dominant.

Any time a .terraformignore file containing a negation was evaluated,
these default rules would be corrected by the parsing algorithm. This
is why this bug was not caught by the tests. So I added another Pack
test that relies only on the default rules.