Optimize order of default rules
By moving the .git/ exclusion rule to the end of the default list,
it's possible to skip the entire .git/ directory walk because no
exclusions appear after it. This should slightly speed up packing
performance when the default rules are used or when no additional
exclusion rules are set by .terraformignore.