fix: header entries representing root dir error
Some tar utilities, most notably Python's built-in TarFile.add,
include a header entry to represent the root of the directory
being tarred and also append a slash to header entries representing
directories. This means the root directory gets the unfortunate name
of "/" which `root.MkdirAll("/")` interprets as an absolute path
and errors.

Since the root directory already exists at this point, we can
just skip the call to create it altogether.